Extract routes to AkkaRoutes

+++ core/src/main/scala/com/ebarrientos/Data.scala
@@ -2,4 +2,4 @@ package com.ebarrientos
 
 case class Address(street: String, zip: String)
 
-case class Data(id: BigDecimal, name: String, address: Address)
+case class Data(id: BigInt, name: String, address: Address)
