+++ core/src/main/scala/com/ebarrientos/DataDao.scala
@@ -3,7 +3,7 @@ package com.ebarrientos
 import zio.Task
 
 trait DataDao {
-  def getOne(id: BigDecimal): Task[Data]
+  def getOne(id: BigInt): Task[Data]
 
   def getList(n: Int): Task[Seq[Data]]
 }
